Fix grammar and sentences
Some sentences are hard to understand, like this one under "Features":

`when you run build, it will copy the files configured in the `pages` options from the dist's sub-folder to dist folder, and then delete the html file` (which is already my correction). 

What do you mean by "copy ... from the dist's sub-folder [...]"? Shouldn't this be the "src's sub-folder, which is copied to the dist folder?
And what do you mean by "[...] and then delete the [rest] html file"? 

Same thing here:
```
+ build mode: it will copy the html file(s) set in this config to the `dist` folder and delete the (useless) folder which stores the html file.
```

Which folders are deleted???